Made BRP New Character Sheet Menu.
Moving away from the Hobby Skills problem and trying a different approach. Started reworking the menus and got the BRP menu up but not complete (Only option is to exit back to the New Sheet Menu). Going to clean up the menus quickly, then take on the Character gen again with a free build style. A "Create Character, Free-Style" option is in the new menu. Needs more with to get the Random Character working with the new menu option I made.