Cleaned Up diceroller
Redid all the dice roll methods but now I'm starting a general die roller which takes arguments for the minimum and maximum random number results. I also started a multi-dice rolling method which takes an additional argument for how many rolls to make. Just need to finish the vector and output of the results. I also need to figure out how to be able to utilize the multi-dice method for things like 3D6 in the BRP_Char.

I also tweaked some parts in BRP_Char;
1. Spaced out the for loop in ProfessionSkillSet()
2. Tested what happens if Hired > 0, it hangs. I believe it gets stuck trying to assign points from a string array that doesn't exists. Until I fix/change this set up I'd need to comment out the ProfessionSkillSet() in FullRandom(). (haven't tested that but it doesn't really matter at this point)