Cleaned Up End of Day Work
Finished what I was working on earlier today. Just added a ; at the end of the cout text results for the many dice roller. I might go back and adjust it to clear the text above after input and output a summary of what was entreated. Maybe even have a sum total result.

I also adjusted the BRP_Char to allow full random for all profession again. I commented out the PickJobSkills() and ProfessionSkillSet() in fullrandom(). Don't really need them for now if I'm not working on Professions for the time being. Better to show off a wider range of characters then a detailed individual artist.