COMPLETED pRNG UPDATE AND BUG FIX
Finally figured out why memory leaks were occurring. In Professions(), when Hired was called is was set to give a random number between 0 and 44. There are 44 total professions but the elements in the array are between 0 and 43. Just changed the roll to be between 0 and 43. Tested and can generate characters without any memory leaks!
This branch can be integrated into the earlier branch. Might have to do that at home...