Started Reworking void HeightandWeight(int z)
Commented out old code and began reworking the method.
By default it takes SIZ  as an input for a switch and rolls random height and weight in cm and kg.
Currently trying to figure out how to output the value in a the following format;
0.00m
0.00g
Need to figure out basic conversion and using doubles.
Later, I will need to convert the cm and kg into imperial in and lbs. Also have a text range default as well.