Finished HeightandWeight method
I updated my roadmap in main.cpp.
Redone HeightandWeight(). It now takes an int argument (for SIZ) and rolls up random values for height in cm and weight in kg. Then, based on settings, will output a string of metric, imperial, or just a basic text description based on the BMI of the random rolls. Maybe overkill but it was fun.
Ready to merge back into Professional_Input stem.