MORE FORMATTING
Mid-day update to formatting the console and was able to make console output be outputted to txt file.
Currently there is an issue where there is some misalignment in the txt file and I have no idea why. It is not due to tabbing like I thought. Will have to research this later.